getListenerをbaseで
